// src/Type/validation/changePasswordValType.type.ts
export type changePasswordValType = {
    valSignInUserNameRequir: string;
    ValSignInUserNameMin: string;
    ValSignInUserNameMax: string;
    valCurrentPassword: string;
    valCurrentPassworRequir: string;
    ValCurrentPassworMin: string;
    ValCurrentPassworMax: string;
    valSignInPasswordRequir: string;
    ValSignInPasswordMin: string;
    ValSignInPasswordMax: string;
    valSignInConfirmPasswordRequir: string;
    ValSignInConfirmPasswordMin: string;
    ValSignInConfirmPasswordMax: string;
    ValSignInConfirmPasswordNotMatch: string;
    valPasswordChangeMessage: string;
};